requirements.txt
pyaudio
websocket-client
numpy
soundfile
python-dotenv
openai
pydub
faiss-cpu
sentence_transformers